checking refresh token expiration

+++ backend/server.js
@@ -28,7 +28,7 @@ const corsOptions = {
   allowedHeaders: ["Content-type", "Authorization", "Access-Control-Allow-Credentials", "Access-Control-Allow-Origin"],
 };
 const app = express();
-
+// console.log(randomBytes(32).toString("hex"))
 
 const PORT = process.env.PORT || 3004;
 const encryptPassword = (password, salt) => {
@@ -55,43 +55,72 @@ async function checkUserExists(username, email) {
   return rows.length > 0;
 }
 
-async function generateRefreshToken(userId) {
-  const refreshToken = randomBytes(64).toString('hex');
-  const expiresAt = new Date(Date.now() + 30 * 24 * 60 * 60 * 1000) // 30 days;
-
+async function generateRefreshToken(user) {
+  // const expiresAt = new Date(Date.now() + 30 * 24 * 60 * 60 * 1000);
+  const expiresAt = new Date(Date.now() +  15 * 1000);
+  // const refreshToken = jwt.sign({ id: user.id, username: user.username }, process.env.REFRESH_SECRET, { expiresIn: "30d" });
+  const refreshToken = jwt.sign({ id: user.id, username: user.username }, process.env.REFRESH_SECRET, { expiresIn: "15s" });
+  const hashToken = hashPassword(refreshToken);
   const insertQuery = 'INSERT INTO refresh_tokens(user_id, token, expires_at) VALUES (?, ?, ?)';
-  const [results, error] = await poolPromise.query({ sql: insertQuery, values: [userId, refreshToken, expiresAt] });
-  return refreshToken;
-}
-async function refreshAccessToken(refreshToken) {
-  const selectQuery = ' SELECT * FROM refresh_tokens WHERE token = ? AND expires_at > NOW()';
-  const [rows] = await poolPromise.query({ sql: selectQuery, values: [refreshToken] })
   
-  const tokenData = rows[0];
-  if (!tokenData) { 
-    console.error('refresh token expired');
+  try {
+    const [results, error] = await poolPromise.query({ sql: insertQuery, values: [user.id, hashToken, expiresAt] });
+  } catch (error) {
+    console.error(`generateRefreshToken: ${error}`);
   }
-  const userSelectQuery = 'SELECT * FROM users WHERE id = ?';
-  const [userRows] = await poolPromise.query({ sql: userSelectQuery, values: [tokenData.user_id] });
-  const user = userRows[0];
-  const newAccessToken = jwt.sign({id: user.id, username: user.username}, process.env.JWT_SECRET, {expiresIn: '1h'})
-  return newAccessToken;
+
+  return refreshToken;
+  // return jwt.sign({ id: user.id, username: user.username }, process.env.REFRESH_SECRET, { expiresIn: "30d" });
 }
-async function checkUserToken(userId, token) {
-  const selectQuery = 'SELECT * FROM refresh_tokens WHERE user_id = ? AND token = ? AND expires_at > NOW()';
-  const [rows] = await poolPromise.query({ sql: selectQuery, values: [userId, token] });
+async function generateAccessToken(user) {
+  return jwt.sign({ id: user.id, username: user.username }, process.env.JWT_SECRET, { expiresIn: '20s' });
+}
+// async function generateRefreshToken(userId) {
+//   const refreshToken = randomBytes(64).toString('hex');
+//   const expiresAt = new Date(Date.now() + 30 * 24 * 60 * 60 * 1000) // 30 days;
+
+//   const insertQuery = 'INSERT INTO refresh_tokens(user_id, token, expires_at) VALUES (?, ?, ?)';
+//   const [results, error] = await poolPromise.query({ sql: insertQuery, values: [userId, refreshToken, expiresAt] });
+//   return refreshToken;
+// }
+// async function refreshAccessToken(refreshToken) {
+//   const selectQuery = ' SELECT * FROM refresh_tokens WHERE token = ? AND expires_at > NOW()';
+//   const [rows] = await poolPromise.query({ sql: selectQuery, values: [refreshToken] })
+  
+//   const tokenData = rows[0];
+//   if (!tokenData) { 
+//     console.error('refresh token expired');
+
+//   }
+//   const userSelectQuery = 'SELECT * FROM users WHERE id = ?';
+//   const [userRows] = await poolPromise.query({ sql: userSelectQuery, values: [tokenData.user_id] });
+//   const user = userRows[0];
+//   const newAccessToken = jwt.sign({id: user.id, username: user.username}, process.env.JWT_SECRET, {expiresIn: '1h'})
+//   return newAccessToken;
+// }
+async function checkUserToken(token) {
+  const hashToken = hashPassword(token);
+  const selectQuery = 'SELECT * FROM refresh_tokens WHERE token = ? AND expires_at > NOW()';
+  const [rows] = await poolPromise.query({ sql: selectQuery, values: [ hashToken] });
 
   return (rows.length > 0);
 }
-const extractToken = (req, res, next) => {
+const extractToken = async (req, res, next) => {
   const authHeader = req.headers['authorization'];
+  // console.log(authHeader);
   if (authHeader && authHeader.startsWith('Bearer ')) {
     const token = authHeader.split(' ')[1];
-    req.token = token;
+    // console.log(!token)
+    // if (!token) return res.sendStauts(403);
+    // const token  = jwt.verify(tokenStr, process.env.JWT_SECRET);
+    req.token =  token == 'null' ? null : token;
+    // console.log(token)
+    // if (!user) return res.sendStauts(401);
+    // req.user = user
+    next();
   } else {
-    req.token = null;
+    req.token = undefined;
   }
-  next();
 }
  
 const pool = mysql.createPool({
@@ -117,7 +146,7 @@ const poolPromise = pool.promise();
 
 app.use(cors(corsOptions));
 
-app.use(extractToken);
+// app.use(extractToken);
 app.use(cookieParser(process.env.COOKIE_SECRET));
 
 app.options("*", cors(corsOptions));
@@ -171,16 +200,59 @@ async function setUpDb() {
 }
 // setUpDb();
 
-app.get("/getInvoices", async (req, res) => {
+async function checkTokens(req, res) {
+  const refreshToken = req.signedCookies['refresh_token'];
   const token = req.token;
+  if (!token) {
+    return 403;
+  };
+  if (!refreshToken) {
+    return 403;
+  }
+  try {
+    const user = jwt.verify(refreshToken, process.env.REFRESH_SECRET);
+    // console.log(refreshToken);
+    const userForAccessToken = jwt.verify(token, process.env.JWT_SECRET);
+    // if (!user || !userForAccessToken) {
+      // console.log(user, "here")
+      // return 403;
+    // }
+    
+  } catch (error) {
+    if (error.name === 'TokenExpiredError') {
+      const deleteQuery = 'DELETE FROM refresh_tokens WHERE token = ? AND expires_at < NOW()';
+      const hashToken = hashPassword(refreshToken);
+      const [result] = await poolPromise.query({ sql: deleteQuery, values: [hashToken] });
+      // console.log(result['affectedRows']);
+      if (result['affectedRows'] > 0) {
+          res.clearCookie('refresh_token', {
+              httpOnly: true,
+              signed: true,
+              secure: true,
+              sameSite: 'None',
+              path: '/' ,
+          });
+      }
+      // if (result)
+      // console.log(hashToken);
+    console.error(`check tokens ${error}`);
+  } else {
+    console.error(`check tokens  ${error}`);
+    }
+    
+    return 403;
+  }
 
-  // console.log(token);
-  // console.log(req.signedCookies['refresh_token']);
-  // console.log(req.cookies);
-  // console.log(req.cookies['refresh_token']);
-  // if (!token) {
-  //   return res.status(401).json({message : "Access Denied"});
-  // }
+  return 200;
+}
+
+app.get("/getInvoices", extractToken, async (req, res) => {
+  
+  const statusCode = await checkTokens(req, res);
+  // console.log(statusCode);
+  if (statusCode === 403) {
+    return res.status(statusCode).json({ message: 'tokens are invalid' });
+  }
   try {
     const selectQuery = "SELECT * FROM invoices";
     const [results] = await poolPromise.query(selectQuery);
@@ -189,7 +261,11 @@ app.get("/getInvoices", async (req, res) => {
     console.error(`getInvoices: ${error}`);
   }
 });
-app.get('/getInvoice/:id', async (req, res) => {
+app.get('/getInvoice/:id', extractToken,  async (req, res) => {
+  const statusCode = await checkTokens(req, res) 
+  if (statusCode === 403) {
+    return res.status(statusCode).json({ message: 'tokens are invalid' });
+  }
   try {
     const { id } = req.params;
     const selectQuery = 'SELECT * FROM invoices WHERE id = ? LIMIT 1';
@@ -199,7 +275,11 @@ app.get('/getInvoice/:id', async (req, res) => {
     console.log(`getInvoice: ${error}`);
   }
 });
-app.post('/saveInvoice', async (req, res) => {
+app.post('/saveInvoice', extractToken, async (req, res) => {
+  const statusCode = await checkTokens(req, res) 
+  if (statusCode === 403) {
+    return res.status(statusCode).json({ message: 'tokens are invalid' });
+  }
   try {
     const {
       id,
@@ -236,7 +316,11 @@ app.post('/saveInvoice', async (req, res) => {
     console.log(`saveInvoice: ${error}`);
   }
 });
-app.put('/updateStatus/:id', async (req, res) => {
+app.put('/updateStatus/:id', extractToken, async (req, res) => {
+  const statusCode = await checkTokens(req, res) 
+  if (statusCode === 403) {
+    return res.status(statusCode).json({ message: 'tokens are invalid' });
+  }
   try {
     const { id } = req.params;
     const { status } = req.body;
@@ -248,7 +332,11 @@ app.put('/updateStatus/:id', async (req, res) => {
     console.error(`updateStatus: ${error}`)
   }
 });
-app.post('/updateInvoice/:id',async (req, res) => {
+app.post('/updateInvoice/:id', extractToken, async (req, res) => {
+  const statusCode = await checkTokens(req, res) 
+  if (statusCode === 403) {
+    return res.status(statusCode).json({ message: 'tokens are invalid' });
+  }
   try {
     const { id } = req.params;
      const {
@@ -303,16 +391,24 @@ app.post('/loginUser', async (req, res) => {
     if (!matchPassword(password, user.password_hash)) {
       return res.status(401).json({ message: "Password is incorrect" });
     }
-
-    const refreshToken = await generateRefreshToken(user.id);
-    const accessToken = await refreshAccessToken(refreshToken);
+    const refreshToken = await generateRefreshToken(user);
+    const accessToken = await generateAccessToken(user);
+  
+    // res.cookie('refresh_token', refreshToken, {
+    //   httpOnly: true,
+    //   signed: true,
+    //   secure: true,
+    //   sameSite: 'None',
+    //   path: '/' ,
+    //   maxAge: 30 * 24 * 60 * 60 * 1000,
+    // });
     res.cookie('refresh_token', refreshToken, {
       httpOnly: true,
       signed: true,
       secure: true,
       sameSite: 'None',
       path: '/' ,
-      maxAge: 30 * 24 * 60 * 60 * 1000,
+      maxAge: 31 * 24 * 60 * 60 * 1000,
     });
     
     res.json({ accessToken });
@@ -324,10 +420,20 @@ app.post('/loginUser', async (req, res) => {
 })
 
 app.post('/logout', (req, res) => {
-  res.clearCookie('refresh_token');
+  res.clearCookie('refresh_token', {
+    httpOnly: true,
+    signed: true,
+    secure: true,
+    sameSite: 'None',
+    path: '/' ,
+  });
   res.send("Logged out successfully");
 })
-app.delete('/deleteInvoice/:id', async (req, res) => {
+app.delete('/deleteInvoice/:id', extractToken, async (req, res) => {
+  const statusCode = await checkTokens(req, res) 
+  if (statusCode === 403) {
+    return res.status(statusCode).json({ message: 'tokens are invalid' });
+  }
   try {
     const { id } = req.params;
     const deleteQuery = 'DELETE FROM invoices WHERE id = ? LIMIT 1';
@@ -347,6 +453,36 @@ app.get("/health-check", async (req, res) => {
   } catch (error) {
     console.error(`health-check: ${error}`);
   }
+});
+
+app.post('/refresh-token', async (req, res) => {
+  const refreshToken = req.signedCookies['refresh_token'];
+  if (!refreshToken) {
+    // console.log(refreshToken);
+    return res.status(403).json({ message: 'refresh token not found, pleas log in agian' });
+  }
+
+  // if (!checkUserToken(refreshToken)) {
+  //   const deleteQuery = 'DELETE FROM refresh_tokens WHERE token = ?';
+  //   const hashToken = hashPassword(refreshToken);
+  //   const [result] = await poolPromise.query({ sql: deleteQuery, values: [hashToken] });
+  //   return res.status(403).json({ message: 'refresh token expired' });
+  // }
+
+  try {
+    const user = jwt.verify(refreshToken, process.env.REFRESH_SECRET);
+    const newAccessToken = await generateAccessToken(user);
+    res.json({ accessToken: newAccessToken });
+  } catch (err) {
+    if (err.name === 'TokenExpiredError') {
+      console.error(`refresh_token expires: ${err.expiresAt}`);
+    } else {
+      console.error(`refresh_token message: ${err.message}`);
+    }
+    res.status(403).send('Invalid refresh token');
+  }
+  // jwt.verify(refreshToken, process.env.)
+
 });
 
 app.listen(PORT, () => {
